fix: displaying menu option data issue
애뮬레이터에서는 다이얼로그가 두 번 띄워지는 이슈 추가 존재
