+++ feature/main/src/main/java/org/swm/att/home/adapter/CategoryMenuAdapter.kt
@@ -1,10 +1,8 @@
 package org.swm.att.home.adapter
 
-import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.FragmentActivity
 import androidx.recyclerview.widget.ListAdapter
 import org.swm.att.common_ui.util.ItemDiffCallback
 import org.swm.att.common_ui.util.ItemTouchHelperListener
@@ -13,11 +11,9 @@ import org.swm.att.domain.entity.response.MenuVO
 import org.swm.att.home.R
 import org.swm.att.home.home.HomeViewModel
 import org.swm.att.home.home.menu.MenuViewHolder
-import org.swm.att.home.home.option.MenuOptionDialog
 
 class CategoryMenuAdapter(
-    private val homeViewModel: HomeViewModel,
-    private val parentContext: Context
+    private val homeViewModel: HomeViewModel
 ) : ListAdapter<MenuVO, MenuViewHolder>(
     ItemDiffCallback<MenuVO>(
         onItemsTheSame = { old, new -> old.id == new.id },
@@ -42,11 +38,6 @@ class CategoryMenuAdapter(
         holder.bind(menu)
         holder.itemView.setOnClickListener {
             homeViewModel.getMenuInfo(menu.id)
-            val menuOptionDialog = MenuOptionDialog(homeViewModel)
-            menuOptionDialog.show(
-                (parentContext as FragmentActivity).supportFragmentManager,
-                MenuOptionDialog::class.java.simpleName
-            )
         }
 //        holder.itemView.setOnTouchListener { _, motionEvent ->
 //            if (motionEvent.actionMasked == android.view.MotionEvent.ACTION_DOWN) {
@@ -63,7 +54,6 @@ class CategoryMenuAdapter(
     override fun onItemMove(fromPosition: Int, toPosition: Int): Boolean {
         val item = getItem(fromPosition)
         val currentList = currentList.toMutableList()
-
         currentList.removeAt(fromPosition)
         currentList.add(toPosition, item)
         submitList(currentList)
