+++ feature/main/src/main/java/org/swm/att/home/home/option/MenuOptionDialog.kt
@@ -2,15 +2,9 @@ package org.swm.att.home.home.option
 
 import android.os.Bundle
 import android.view.View
-import android.widget.Toast
 import androidx.fragment.app.viewModels
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.lifecycleScope
-import androidx.lifecycle.repeatOnLifecycle
 import androidx.recyclerview.widget.LinearLayoutManager
-import kotlinx.coroutines.launch
 import org.swm.att.common_ui.base.BaseDialog
-import org.swm.att.common_ui.util.state.UiState
 import org.swm.att.domain.entity.request.OrderedMenuVO
 import org.swm.att.domain.entity.response.MenuWithRecipeVO
 import org.swm.att.home.R
@@ -19,18 +13,22 @@ import org.swm.att.home.databinding.DialogMenuOptionBinding
 import org.swm.att.home.home.HomeViewModel
 
 class MenuOptionDialog(
-    private val homeViewModel: HomeViewModel
+    private val homeViewModel: HomeViewModel,
+    private val menu: MenuWithRecipeVO
 ): BaseDialog<DialogMenuOptionBinding>(R.layout.dialog_menu_option) {
     private lateinit var menuOptionAdapter: MenuOptionAdapter
     private val menuOptionViewModel by viewModels<MenuOptionViewModel>()
-    private lateinit var menu: MenuWithRecipeVO
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         setDialogCloseBtnListener()
         initOptionsRecyclerView()
         setOptionAddBtnClickListener()
-        setSelectedMenuInfoObserver()
+        setMenuOptionData()
+    }
+
+    private fun setMenuOptionData() {
+        menuOptionAdapter.submitList(menu.option)
     }
 
     private fun setDialogCloseBtnListener() {
@@ -66,32 +64,6 @@ class MenuOptionDialog(
             )
             homeViewModel.addSelectedMenu(selectedMenuWithOptions)
             dismiss()
-        }
-    }
-
-    private fun setSelectedMenuInfoObserver() {
-        lifecycleScope.launch {
-            repeatOnLifecycle(Lifecycle.State.STARTED) {
-                homeViewModel.getMenuInfoState.collect { uiState ->
-                    when (uiState) {
-                        is UiState.Loading -> {/* nothing */
-                        }
-
-                        is UiState.Success -> {
-                            uiState.data?.let {
-                                menu = it
-                                menuOptionAdapter.submitList(it.option)
-                            }
-                        }
-
-                        is UiState.Error -> {
-                            Toast.makeText(requireContext(), uiState.errorMsg, Toast.LENGTH_SHORT)
-                                .show()
-                            dismiss()
-                        }
-                    }
-                }
-            }
         }
     }
